requirements.txt
Pillow==9.4.0
PySimpleGUI==4.60.4